requirements.txt
openai>=1.0.0
anthropic>=0.3.0
python-dotenv>=1.0.0
requests>=2.0.0
pytest>=7.0.0
httpx>=0.23.0
pydantic>=2.11.0
typing-extensions>=4.13.0


# Optional
black
flake8
